Create p2p-vs-bridge-views.sql
This is just a query for now, but we can build on it and potentially make it a database view. Alternatively we can store the query directly in the AWS Redshift query editor but I think having the queries in Git has the advantage of... well, version control and change tracking.