feat: Makes it possible to mass update the test results.
This commit makes it possible to mass update the test results. The use-case is
the following:

Suppose, for example, that you have modified the feature extractor to add a
`startDate` field in the track features. When you run the tests they fail and
they present you with the diff. After you have reviewed the diff and you are
satisfied with the result you will need to update all the JSON results
accordingly so that the Jest tests actually pass. With this commit you can
simply run `npm run test:fix` to do that last part.

As a bonus point, this procedure will enforce uniform formatting rules for our
JSON result files. If you notice currently we have some with 2-space
indentation and others with 4-space indentation.

The way `npm run test:fix` is that it will generate a stringified JSON with a
special replacer that garantees that properties/fields of the stringified
object will be sorted so the stringified representation is truly stable.

Note that I have already run `npm run test:fix` which has re-formatted the JSON
result files but their contained hasn't changed.